final conditional support added to root setup file
