+++ skimage/future/detect/cascade.pyx
@@ -5,6 +5,7 @@
 # distutils: language = c++
 
 
+
 import numpy as np
 cimport numpy as cnp
 cimport safe_openmp as openmp
