+++ skimage/future/detect/setup.py
@@ -3,61 +3,8 @@
 from __future__ import print_function
 from skimage._build import cython
 import os
-import tempfile
-import shutil
-from numpy.distutils.command.build_ext import build_ext
-from distutils.errors import CompileError, LinkError
 
 base_path = os.path.abspath(os.path.dirname(__file__))
-
-compile_flags = ['-fopenmp']
-link_flags = ['-fopenmp']
-
-code = """#include <omp.h>
-int main(int argc, char** argv) { return(0); }"""
-
-class Checker(build_ext):
-
-    def can_compile_link(self):
-
-        cc = self.compiler
-        fname = 'test.c'
-        cwd = os.getcwd()
-        tmpdir = tempfile.mkdtemp()
-
-        try:
-            os.chdir(tmpdir)
-            with open(fname, 'wt') as fobj:
-                fobj.write(code)
-            try:
-                objects = cc.compile([fname],
-                                     extra_postargs=compile_flags)
-            except CompileError:
-                return False
-            try:
-                # Link shared lib rather then executable to avoid
-                # http://bugs.python.org/issue4431 with MSVC 10+
-                cc.link_shared_lib(objects, "testlib",
-                                   extra_postargs=link_flags)
-            except (LinkError, TypeError):
-                return False
-        finally:
-            os.chdir(cwd)
-            shutil.rmtree(tmpdir)
-        return True
-
-    def build_extensions(self):
-        """ Hook into extension building to check compiler flags """
-
-        print('-----------------------')
-
-        if self.can_compile_link():
-
-            for ext in self.extensions:
-                ext.extra_compile_args += compile_flags
-                ext.extra_link_args += link_flags
-
-        build_ext.build_extensions(self)
 
 
 def configuration(parent_package='', top_path=None):
@@ -75,19 +22,15 @@ def configuration(parent_package='', top_path=None):
                          language="c++")
     return config
 
-cmdclass = dict(build_ext=Checker)
-
 if __name__ == '__main__':
     from numpy.distutils.core import setup
 
     conf = configuration(top_path='').todict()
-
-    conf['cmdclass'] = cmdclass
 
     setup(maintainer='scikit-image Developers',
           maintainer_email='scikit-image@googlegroups.com',
           description='Object detection framework',
           url='https://github.com/scikit-image/scikit-image',
           license='Modified BSD',
-          **(conf)
+          **conf
           )
